+++ src/addons/display_commander/settings/advanced_tab_settings.cpp
@@ -51,6 +51,7 @@ AdvancedTabSettings::AdvancedTabSettings()
       dlls_to_load_before("DllsToLoadBefore", "", "DisplayCommander"),
       fake_nvapi_enabled("FakeNvapiEnabled", false, "DisplayCommander"),
       suppress_minhook("SuppressMinhook", false, "DisplayCommander.Safemode"),
+      suppress_wgi_globally("SuppressWgiGlobally", false, "DisplayCommander"),
       suppress_wgi_enabled("SuppressWgiEnabled", false, "DisplayCommander"),
       suppress_wgi_for_unity("SuppressWgiForUnity", false, "DisplayCommander"),
       suppress_wgi_for_non_unity_games("SuppressWgiForNonUnityGames", false, "DisplayCommander"),
@@ -93,6 +94,7 @@ void AdvancedTabSettings::SaveAll() {
     safemode.Save();
     fake_nvapi_enabled.Save();
     suppress_minhook.Save();
+    suppress_wgi_globally.Save();
     suppress_wgi_enabled.Save();
     suppress_wgi_for_unity.Save();
     suppress_wgi_for_non_unity_games.Save();
@@ -130,13 +132,14 @@ std::vector<ui::new_ui::SettingBase*> AdvancedTabSettings::GetAllSettings() {
         &enable_timeslowdown_shortcut, &enable_adhd_toggle_shortcut, &enable_autoclick_shortcut,
         &enable_input_blocking_shortcut, &enable_display_commander_ui_shortcut, &enable_performance_overlay_shortcut,
         &safemode, &dll_loading_delay_ms, &dlls_to_load_before, &fake_nvapi_enabled, &suppress_minhook,
-        &suppress_wgi_enabled, &suppress_wgi_for_unity, &suppress_wgi_for_non_unity_games, &debug_layer_enabled,
-        &debug_break_on_severity,
-        &auto_hide_discord_overlay, &suppress_window_changes, &win_up_grace_seconds, &enable_presentmon_tracing,
-        &presentmon_provider_dxgkrnl, &presentmon_provider_dxgi, &presentmon_provider_dwm, &presentmon_provider_d3d9,
-        &disable_dpi_scaling, &enable_dx11_vtable_hooks, &texture_tracking_enabled, &d3d11_texture_caching_enabled,
-        &d3d11_texture_caching_1d_enabled, &d3d11_texture_caching_3d_enabled, &texture_cache_content_hash_cap_kb,
-        &dump_textures_enabled, &show_steam_achievement_notifications, &show_steam_achievement_counter_increased};
+        &suppress_wgi_globally, &suppress_wgi_enabled, &suppress_wgi_for_unity, &suppress_wgi_for_non_unity_games,
+        &debug_layer_enabled,
+        &debug_break_on_severity, &auto_hide_discord_overlay, &suppress_window_changes, &win_up_grace_seconds,
+        &enable_presentmon_tracing, &presentmon_provider_dxgkrnl, &presentmon_provider_dxgi, &presentmon_provider_dwm,
+        &presentmon_provider_d3d9, &disable_dpi_scaling, &enable_dx11_vtable_hooks, &texture_tracking_enabled,
+        &d3d11_texture_caching_enabled, &d3d11_texture_caching_1d_enabled, &d3d11_texture_caching_3d_enabled,
+        &texture_cache_content_hash_cap_kb, &dump_textures_enabled, &show_steam_achievement_notifications,
+        &show_steam_achievement_counter_increased};
 }
 
 }  // namespace settings